// web/gen/clob.ts
export interface OrderDetail {
    /**
     * @generated from protobuf field: string tx_id = 1
     */
    txId: string;
    /**
     * @generated from protobuf field: string account_id = 2
     */
    accountId: string;
    /**
     * @generated from protobuf field: uint64 limit_yes_price = 6
     */
    limitYesPrice: bigint;
    /**
     * @generated from protobuf field: uint64 shares_remaining = 7
     */
    sharesRemaining: bigint;
    /**
     * @generated from protobuf field: uint32 side = 8
     */
    side: number;
    /**
     * @generated from protobuf field: uint32 action = 9
     */
    action: number;
}
/**
 * @generated from protobuf message clob.BookSnapshot
 */
export interface BookSnapshot {
    /**
     * @generated from protobuf field: repeated clob.OrderDetail bids = 1
     */
    bids: OrderDetail[];
    /**
     * @generated from protobuf field: repeated clob.OrderDetail asks = 2
     */
    asks: OrderDetail[];
}
// ...
class OrderDetail$Type extends MessageType<OrderDetail> {
    constructor() {
        super("clob.OrderDetail", [
            { no: 1, name: "tx_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "account_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 6, name: "limit_yes_price", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 7, name: "shares_remaining", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 8, name: "side", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
            { no: 9, name: "action", kind: "scalar", T: 13 /*ScalarType.UINT32*/ }
        ]);
    }
    create(value?: PartialMessage<OrderDetail>): OrderDetail {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.txId = "";
        message.accountId = "";
        message.limitYesPrice = 0n;
        message.sharesRemaining = 0n;
        message.side = 0;
        message.action = 0;
        if (value !== undefined)
            reflectionMergePartial<OrderDetail>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: OrderDetail): OrderDetail {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string tx_id */ 1:
                    message.txId = reader.string();
                    break;
                case /* string account_id */ 2:
                    message.accountId = reader.string();
                    break;
                case /* uint64 limit_yes_price */ 6:
                    message.limitYesPrice = reader.uint64().toBigInt();
                    break;
                case /* uint64 shares_remaining */ 7:
                    message.sharesRemaining = reader.uint64().toBigInt();
                    break;
                case /* uint32 side */ 8:
                    message.side = reader.uint32();
                    break;
                case /* uint32 action */ 9:
                    message.action = reader.uint32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: OrderDetail, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string tx_id = 1; */
        if (message.txId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.txId);
        /* string account_id = 2; */
        if (message.accountId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.accountId);
        /* uint64 limit_yes_price = 6; */
        if (message.limitYesPrice !== 0n)
            writer.tag(6, WireType.Varint).uint64(message.limitYesPrice);
        /* uint64 shares_remaining = 7; */
        if (message.sharesRemaining !== 0n)
            writer.tag(7, WireType.Varint).uint64(message.sharesRemaining);
        /* uint32 side = 8; */
        if (message.side !== 0)
            writer.tag(8, WireType.Varint).uint32(message.side);
        /* uint32 action = 9; */
        if (message.action !== 0)
            writer.tag(9, WireType.Varint).uint32(message.action);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message clob.OrderDetail
 */
export const OrderDetail = new OrderDetail$Type();
// @generated message type with reflection information, may provide speed optimized methods
class BookSnapshot$Type extends MessageType<BookSnapshot> {
    constructor() {
        super("clob.BookSnapshot", [
            { no: 1, name: "bids", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => OrderDetail },
            { no: 2, name: "asks", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => OrderDetail }
        ]);
    }
    create(value?: PartialMessage<BookSnapshot>): BookSnapshot {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.bids = [];
        message.asks = [];
        if (value !== undefined)
            reflectionMergePartial<BookSnapshot>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: BookSnapshot): BookSnapshot {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated clob.OrderDetail bids */ 1:
                    message.bids.push(OrderDetail.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* repeated clob.OrderDetail asks */ 2:
                    message.asks.push(OrderDetail.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: BookSnapshot, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated clob.OrderDetail bids = 1; */
        for (let i = 0; i < message.bids.length; i++)
            OrderDetail.internalBinaryWrite(message.bids[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* repeated clob.OrderDetail asks = 2; */
        for (let i = 0; i < message.asks.length; i++)
            OrderDetail.internalBinaryWrite(message.asks[i], writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message clob.BookSnapshot
 */
export const BookSnapshot = new BookSnapshot$Type();